refactor: split webview message logic into managers
